requirements.txt
python-telegram-bot==20.3
google-generativeai>=0.3.2
python-dotenv
Pillow